+++ CHANGELOG.md
@@ -12,9 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 * Changed topological_sort as method of DirectedGraph class
+* Changed backdoor adjustement sets to all adjustment sets and minimal adjustment sets
 
 ### Removed
 ### Fixed
+* Fixed all simple paths return type
 
 ## [0.0.1] - 2020-01-06
 ### Added